Restore timer_subscribe and delete hello example
